% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster_and_summarize_prediction_explanations}
\alias{cluster_and_summarize_prediction_explanations}
\title{Cluster and Summarize Prediction Explanations}
\usage{
cluster_and_summarize_prediction_explanations(model_object, scoring_df,
  num_neighbors = 30, min_dist = 10^-100, min_points = 150,
  max_explanations = 3, num_feature_summarizations = 10)
}
\arguments{
\item{model_object}{The \code{dataRobotModel} from which to source prediction explanations}

\item{scoring_df}{The DataFrame on which prediction explanations will be calculated}

\item{num_neighbors}{The number of neighbors argument for dimensionality reduction (see \code{n_neighbors} argument in \code{\link[umap]{umap.defaults}})}

\item{min_dist}{The minimum distance argument for dimensionality reduction (see \code{min_dist} argument in \code{\link[umap]{umap.defaults}})}

\item{min_points}{The minimum points parameter for clustering (see \code{minPts} argument in \code{\link[dbscan]{hdbscan}})}

\item{max_explanations}{The maximum number of prediction explanations to retrieve for each observation.}

\item{num_feature_summarizations}{How many features to include in the summarization of clusters (ordered by feature importance).}
}
\value{
Returns an object of class \code{dataRobotPEClusterResults}, which has
  \code{\link[=summary.dataRobotPEClusterResults]{summary}} and
  \code{\link[=plot.dataRobotPEClusterResults]{plot}} methods. An object of
  this class is a list with the following components: \itemize{
  \item
  \code{plot_data}. A data frame containing the cluster IDs for each point
  from \code{scoring_df}, along with their location in the reduced prediction
  explanation space. See \code{\link{plot.dataRobotPEClusterResults}}.
  \item \code{summary_data}. A data frame containing one row per cluster,
  along with summarizations of the top features. See
  \code{\link{summary.dataRobotPEClusterResults}}.
  \item \code{cluster_ids} A
  vector of cluster IDs for each row from \code{scoring_df}.
  \item \code{strength_matrix}: A data frame of the prediction
  explanation strengths of each feature, for each row from \code{scoring_df}.
  \item \code{pe_frame}: The data frame returned by
  \code{\link[datarobot]{GetPredictionExplanationsRowsAsDataFrame}}
 }
}
\description{
Runs dimensionality reduction and clustering on prediction explanations, returning a dataset
suitable for visualization, a cluster-level summarization, and detailed results.
}
