% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datarobot.pe.clustering.R
\docType{package}
\name{datarobot.pe.clustering-package}
\alias{datarobot.pe.clustering-package}
\title{datarobot.pe.clustering: prediction explanation clustering for DataRobot}
\description{
A package for clustering row-level prediction explanations from a machine learning model in the DataRobot platform.
}
\details{
The primary function to be invoked for the prediction explanation clustering pipeline is \code{\link{cluster_and_summarize_prediction_explanations}}.
}
