% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{dataRobotPEClusterResults}
\alias{dataRobotPEClusterResults}
\title{Constructor for dataRobotPEClusterResults S3 class}
\usage{
dataRobotPEClusterResults(...)
}
\arguments{
\item{...}{Results data to package together}
}
\description{
Constructor for dataRobotPEClusterResults S3 class. This constructor packages up provided data into a list with assigned class \code{dataRobotPEClusterResults}
}
