% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{is.dataRobotPEClusterResults}
\alias{is.dataRobotPEClusterResults}
\title{Check for dataRobotPEClusterResults S3 class}
\usage{
is.dataRobotPEClusterResults(object)
}
\arguments{
\item{object}{any \code{R} object}
}
\description{
Function to check if an object is a dataRobotPEClusterResults object.
}
