% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{plot.dataRobotPEClusterResults}
\alias{plot.dataRobotPEClusterResults}
\title{Plot method for S3 objects of class dataRobotPEClusterResults}
\usage{
\method{plot}{dataRobotPEClusterResults}(x, ...)
}
\arguments{
\item{x}{The \code{dataRobotPEClusterResults} to plot}

\item{...}{Additional arguments to pass to \code{plot.default}}
}
\description{
Plots observations on the reduced Prediction Explanation space, coloring by cluster.
}
