% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{summary.dataRobotPEClusterResults}
\alias{summary.dataRobotPEClusterResults}
\title{Summary method for S3 objects of class dataRobotPEClusterResults}
\usage{
\method{summary}{dataRobotPEClusterResults}(object, ...)
}
\arguments{
\item{object}{The \code{dataRobotPEClusterResults} object}

\item{...}{Additional arguments (unused)}
}
\value{
A summary data frame.
}
\description{
Summarizes each Prediction Explanation cluster by aggregating feature values for the cluster members.
}
\details{
Numerical features are summarized by their mean, whereas categorical variables are summarized
by their most frequent value.
}
